Refactored the neighborhood code a lot.
Renamed a bunch of stuff. Moved a method to stage.lisp where it should live.

Removed passing the stage parameter from neighborhood methods
since when making the neighborhood, I just store the stage in the
neighborhood so we can use it when needed.

Neighborhood definition functions are isolated and easy to create now.

Mapping a function over an arbitrary neighborhood exists, but it uses
a slow method by default. You can supply better nh map functions as
needed for different definitions of neighborhoods.

Rewrote the tests.

To see it work, call:

;; no stage clipping
(neighbor-tests 10 10 7 1)

and

;; stage clipping
(neighbor-tests 0 0 7 1)

then look at the output.

There is a TODO that specifies how to create a NH such that there are no
(tile ...) calls to make them almost free to make. You only pay for them
when you actually look something up in them.

Rewrote NREF to be more conservative in the work it is willing to perform.