Half done with API change. Renamed some things, it'll probably break your code axion. Not done with API renaming.
Added more return values to nh-map which returns the results list
like mapcar and also the number of tiles examined while performing
the map-fn. Useful for checking performance of the mapping function.

Wrote a specific nh-map function for the ortho neighborhood. It
dropped the tiles examined from 225 with a distance 7 neighborhood
to 29 examined tiles. :) I amended the test function to display the
num examined tiles the nh-map function performs.

Redid how the mathematical neighborhoods are defined. There is no
longer a lambda closure. Instead I reified the closed information by
hold it in the neighorhood struct and I pass the struct ala single
dispatch. This removed the closure requirement and it looks more
normal wrt OOP programming.

Fixed the testing code to use the new stuff.