Revert "Precompute during the construction of a neighborhood if it needs clipping."
It turns out in tests that this ends up being slower. I'm a little perplexed
why this is true. But meh. Out it goes. It is such a simple change if we
discover why it was slow and make it faster later, we can look it up in the
repo.

This reverts commit 5bb274da05a4e1283cde807cadfc07be3c7ba969.