Rough draft of a feature-set keyword storage mechanism.
This stores feature keywords in a hash table as bit position in an integer.
You can convert sets of feature keywords into a single integer and vice
versa. I do some example bit twiddling to solve an intersection problem.

Run (testme) to see how it works. The testme procedure should probably test
creation and inserting of like 400 autogenerated feature keywords and some
randomly sized intersection problems to see how fast it actually is.
It does save a pile on memory if you store the value version of the keyword
sets.

I'm missing operators like adding a feature keyword to a pre-existing integer,
(analogous to pushnew a feature keyword onto a list) etc, etc, etc.

Hey, this was a rough draft. :)